+++ build.sbt
@@ -19,7 +19,7 @@ val common = Seq(
   )
 )
 
-lazy val core = (project in file("."))
+lazy val core = (project in file("core"))
   .settings(common)
   .settings(
     name := "foperator-core",
@@ -52,14 +52,25 @@ lazy val tests = (project in file("tests"))
   ).dependsOn(core, testkit)
 
 // skuber backend
-lazy val skuber = (project in file("skuber"))
+lazy val skuber = (project in file("backends/skuber"))
   .settings(common)
   .settings(
-    name := "foperator-skuber",
+    name := "foperator-backend-skuber",
     libraryDependencies ++= Seq(
       "io.skuber" %% "skuber" % "2.4.0",
       "co.fs2" %% "fs2-reactive-streams" % fs2Version,
       "com.typesafe.akka" %% "akka-slf4j" % "2.5.29",
+    )
+  ).dependsOn(core)
+
+// kubernetes-client backend
+lazy val kclient = (project in file("backends/kubernetes-client"))
+  .settings(common)
+  .settings(
+    name := "foperator-backend-kubernetes-client",
+    libraryDependencies ++= Seq(
+      // v0.7.0 requires cats 3.x
+      "com.goyeau" %% "kubernetes-client" % "0.6.0"
     )
   ).dependsOn(core)
 
@@ -79,3 +90,5 @@ lazy val sample = (project in file("sample"))
       "mutator-test-live" -> "net.gfxmonk.foperator.sample.MutatorTestLive",
     ),
   ).dependsOn(core, testkit, skuber).enablePlugins(PackPlugin)
+
+lazy val all = (project in file(".")).settings(hiddenProjectSettings).aggregate(testkit, tests, skuber, kclient, sample)
