+++ core/src/main/scala/net/gfxmonk/foperator/Event.scala
