+++ core/src/main/scala/net/gfxmonk/foperator/Id.scala
