+++ core/src/main/scala/net/gfxmonk/foperator/ListOptions.scala
@@ -1,0 +1,12 @@
+package net.gfxmonk.foperator
+
+// TODO less stringly typed...
+case class ListOptions(
+  namespace: String = "default",
+  labelSelector: List[String] = Nil,
+  fieldSelector: List[String] = Nil,
+)
+
+object ListOptions {
+  val all = ListOptions()
+}
