+++ core/src/main/scala/net/gfxmonk/foperator/ReconcileOptions.scala
