+++ core/src/main/scala/net/gfxmonk/foperator/ReconcileSource.scala
