+++ core/src/main/scala/net/gfxmonk/foperator/implicits.scala
