+++ core/src/main/scala/net/gfxmonk/foperator/internal/BackendCompanion.scala
