+++ core/src/main/scala/net/gfxmonk/foperator/internal/Broadcast.scala
