+++ core/src/main/scala/net/gfxmonk/foperator/internal/ErrorCount.scala
