+++ core/src/main/scala/net/gfxmonk/foperator/internal/IOUtil.scala
