+++ core/src/main/scala/net/gfxmonk/foperator/internal/Logging.scala
