+++ core/src/main/scala/net/gfxmonk/foperator/internal/ReconcileLoop.scala
